ref: move verification tests into module
